#16: Converted offset to BigInteger.

+++ core/src/main/java/io/parsingdata/metal/data/ByteStreamSource.java
@@ -34,7 +34,7 @@ public class ByteStreamSource extends Source {
     }
 
     @Override
-    protected byte[] getData(final long offset, final BigInteger length) {
+    protected byte[] getData(final BigInteger offset, final BigInteger length) {
         if (!isAvailable(offset, length)) { throw new IllegalStateException("Data to read is not available ([offset=" + offset + ";length=" + length + ";source=" + this + ")."); }
         try {
             return input.read(offset, length.intValue());
@@ -44,7 +44,7 @@ public class ByteStreamSource extends Source {
     }
 
     @Override
-    protected boolean isAvailable(final long offset, final BigInteger length) {
+    protected boolean isAvailable(final BigInteger offset, final BigInteger length) {
         return input.isAvailable(offset, length.intValue());
     }
